--- location.py
@@ -1,10 +1,0 @@
-class Location():
-    latitude:str   = "0"
-    longitude:str  = "0"
-    city:str       = ""
-    state:str      = ""
-    country:str    = ""
-    error:str      = ""
-    status         = False
-    
-
